Revert live config panel
This fails because iterable object properties are special (proxied) to provide iterable members, and can't easily by duplicated.

It fails because when it is removed from the DOM, the consumers are removed even tho they are *still referenced* (through the common `src.payload`) causing it to terminate early,

This needs a test case and fix in AI-UI to improve the reference counting of iterable propeerty queues.